Revise checking conditions for skipping heap-merge sort
Note number of offset-length lists to be heap-merged is == nprocs_recv.
The computational complexity of heap-merge sort is proportional to not
only the total number of offset-length pairs but also nprocs_recv.
Given the same total number of offset-length pairs to be heap-merged,
the larger number of lists, the more expensive the heap-merge sort.

For example, the total number of offset-length pairs is 200.
* If there are only 2 lists, each of size 100, the cost heap-merge is O(200).
* If there are 10 lists, each of size 20, the cost heap-merge is 9xO(40).

This commit also removes hint 'romio_ds_wr_lb'.