ad_lustre_open.c when striping count is not set, use number of compute nodes
Read the folder's striping parameters if not all striping parameters
are set by the user hints, which include
  * striping_unit (default 0)
  * striping_factor (default 0)
  * start_iodevice (default -1)
  * lustre_overstriping_ratio (default 1)

Add constant PNCIO_LUSTRE_MAX_OSTS of value 256. This is in case the
number of compute nodes is larger than the available number of OSTs.
A better approach is to detect that at the run-time.
However, currently I could not find a way to do so.