Revise strategy of when to reuse user's I/O buffer in MPI-IO/PNCIO calls
Now all get and put APIs are going through the INA subroutines. File
view and buffer view can both be flattened in PnetCDF, which can be
directly used by PNCIO APIs, without constructing any MPI data types.

For writes, the strategy is now changed to use user's buffer when
calling MPI-IO/PNCIO write APIs, only when the following 3 conditions
are fulfilled.
1. the file offset-length pairs and the associated buffer view are not
   re-ordered (i.e. sorted),
2. have no true overlap among offset-length pairs (i.e. buffer size is
   not shrunk), and
3. aggregator's buffer size is equal to the write size (i.e.
   non-aggregators have no data to write/or there is no
   non-aggregators).
These 3 conditions should be the most common case in most of the modern
applications, such as E3SM.

If the above condition cannot be fulfilled, then we allocate a temporary
contiguous buffer, copy the write data over, and use it in calls to
MPI-IO/PNCIO file write APIs.

Reads are the opposite.